Trait pallet_migrations::Migration

pub trait Migration {
    fn friendly_name(&self) -> &str;
    fn migrate(&self, available_weight: Weight) -> Weight;
}

A Migration that must happen on-chain upon a runtime-upgrade

Required Methods

fn friendly_name(&self) -> &str

A human-readable name for this migration. Also used as storage key.

fn migrate(&self, available_weight: Weight) -> Weight

Perform the required migration and return the weight consumed.

Currently there is no way to migrate across blocks, so this method must (1) perform its full migration and (2) not produce a block that has gone over-weight. Not meeting these strict constraints will lead to a bricked chain upon a runtime upgrade because the parachain will not be able to produce a block that the relay chain will accept.