Type Alias pallet_migrations::ActiveCursorOf

pub type ActiveCursorOf<T> = ActiveCursor<RawCursorOf<T>, BlockNumberFor<T>>;

Convenience alias for ActiveCursor.

Aliased Type

struct ActiveCursorOf<T> {
    pub index: u32,
    pub inner_cursor: Option<BoundedVec<u8, <T as Config>::CursorMaxLen>>,
    pub started_at: <<<T as Config>::Block as HeaderProvider>::HeaderT as Header>::Number,
}

Fields
index: u32

The index of the migration in the MBM tuple.

inner_cursor: Option<BoundedVec<u8, <T as Config>::CursorMaxLen>>

The cursor of the migration that is referenced by index.

started_at: <<<T as Config>::Block as HeaderProvider>::HeaderT as Header>::Number

The block number that the migration started at.

This is used to calculate how many blocks it took.